Add styles.css, update: textbox, button, and list
Add styles.css, for now just make body, html have a font family of Open
Sans (Google fonts) and set the height of body to 100%.

In head of Index.html: link styles.css, link Open Sans, and set title of
window.
In body of Index.html: Add paragraph with id of "list", on the button,
update it so text in textbox gets put into list paragraph.